requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic[email]==2.5.3
google-generativeai==0.3.2
httpx>=0.24.0
PyPDF2==3.0.1
python-dotenv==1.0.0
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
asyncpg==0.29.0
sqlmodel>=0.0.8
alembic>=1.10.0
